+++ src/Components/renderBody/index.tsx
@@ -1,0 +1,3 @@
+import { CalculatorBody } from "./CalculatorBody";
+
+export default CalculatorBody;
